Fix Lua SafeEnum and pair conversion
